+++ index.html
@@ -108,5 +108,5 @@
     </div>
   </body>
 
-  <script src="kdb.js"></script>
+  <script src="bootstrap.js"></script>
 </html>
